Clang tidy code cleanup

+++ NodeBrain/src/GAPI/Vulkan/VulkanRendererAPI.cpp
@@ -109,8 +109,8 @@ namespace NodeBrain
 		VkViewport viewport = {};
 		viewport.x = 0.0f;
 		viewport.y = 0.0f;
-		viewport.width = extent.width;
-		viewport.height = extent.height;
+		viewport.width = (float)extent.width;
+		viewport.height = (float)extent.height;
 		viewport.minDepth = 0.0f;
 		viewport.maxDepth = 1.0f;
 		vkCmdSetViewport(m_ActiveCmdBuffer, 0, 1, &viewport);
